Combine test cases for jobstatusdata (#5697)
